### 2022-09-24 #4783 @editorialbot
[PRE REVIEW]: BlenderProc2
**Submitting author:** @themasterlink (<a href="http://orcid.org/0000-0002-1557-2234">Maximilian Denninger</a>)
**Repository:** https://github.com/DLR-RM/BlenderProc
**Branch with paper.md** (empty if default branch):
**Version:** v2.4.1
