### 2022-09-24 #4783 @arfon
@themasterlink – many thanks for your submission to JOSS. I see you note a rewrite of the API for BlenderProc2 – would it be possible to provide a diff between the earlier version of the codebase and this one submitted for review here? e.g., like described here: https://docs.github.com/en/pull-requests/committing-changes-to-your-project/viewing-and-comparing-commits/comparing-commits
