### 2022-09-26 4e5be7b
feat(paper.bib): add DOI
