+++ paper.bib
@@ -16,13 +16,8 @@
   title={The Replica Dataset: A Digital Replica of Indoor Spaces},
   author={Straub, Julian and Whelan, Thomas and Ma, Lingni and Chen, Yufan and Wijmans, Erik and Green, Simon and Engel, Jakob J and Mur-Artal, Raul and Ren, Carl and Verma, Shobhit and others},
   journal={arXiv preprint arXiv:1906.05797},
+  doi = {10.48550/ARXIV.1906.05797},
   year={2019}
-}
-@inproceedings{habitat19iccv,
-  title     =     {Habitat: {A} {P}latform for {E}mbodied {AI} {R}esearch},
-  author    =     {{Manolis Savva*} and {Abhishek Kadian*} and {Oleksandr Maksymets*} and Yili Zhao and Erik Wijmans and Bhavana Jain and Julian Straub and Jia Liu and Vladlen Koltun and Jitendra Malik and Devi Parikh and Dhruv Batra},
-  booktitle =     {Proceedings of the IEEE/CVF International Conference on Computer Vision (ICCV)},
-  year      =     {2019}
 }
 @Article{suncg,
   author      = {Shuran Song and Fisher Yu and Andy Zeng and Angel X. Chang and Manolis Savva and Thomas Funkhouser},
@@ -33,13 +28,14 @@
   eprinttype  = {arXiv},
   abstract    = {This paper focuses on semantic scene completion, a task for producing a complete 3D voxel representation of volumetric occupancy and semantic labels for a scene from a single-view depth map observation. Previous work has considered scene completion and semantic labeling of depth maps separately. However, we observe that these two problems are tightly intertwined. To leverage the coupled nature of these two tasks, we introduce the semantic scene completion network (SSCNet), an end-to-end 3D convolutional network that takes a single depth image as input and simultaneously outputs occupancy and semantic labels for all voxels in the camera view frustum. Our network uses a dilation-based 3D context module to efficiently expand the receptive field and enable 3D context learning. To train our network, we construct SUNCG - a manually created large-scale dataset of synthetic 3D scenes with dense volumetric annotations. Our experiments demonstrate that the joint model outperforms methods addressing each task in isolation and outperforms alternative approaches on the semantic scene completion task.},
   file        = {online:http\://arxiv.org/pdf/1611.08974v1:PDF},
+  doi = {10.48550/ARXIV.1611.08974},
   keywords    = {cs.CV},
-  owner       = {denn_ma},
   timestamp   = {2019.10.11},
 }
 @Article{Sundermeyer2019,
   author      = {Martin Sundermeyer and Zoltan-Csaba Marton and Maximilian Durner and Manuel Brucker and Rudolph Triebel},
   title       = {Implicit 3D Orientation Learning for 6D Object Detection from RGB Images},
+  doi = {10.48550/ARXIV.1902.01275},
   date        = {2019-02-04},
   eprint      = {1902.01275v2},
   eprintclass = {cs.CV},
@@ -47,18 +43,19 @@
[…]
