### 2022-10-05 #4783 @Kevin-Mattheus-Moerman
@themasterlink I had a quick look at the paper. I recommend that you expand the title a bit e.g. in the more usual format: `softwareName: What it does in about one sentence`. You can update the paper here by calling `@editorialbot generate pdf`.
