### 2022-10-24 7063e25
initial sdf writer
Signed-off-by: Mohamad <no.future2e1@gmail.com>
