### 2022-10-27 de6e246
added sdf writer utility
Signed-off-by: Mohamad <mohamad@openrobotics.org>
