+++ blenderproc/api/writer/__init__.py
@@ -1,4 +1,5 @@
 from blenderproc.python.writer.GifWriterUtility import write_gif_animation
 from blenderproc.python.writer.BopWriterUtility import write_bop
 from blenderproc.python.writer.CocoWriterUtility import write_coco_annotations
+from blenderproc.python.writer.WorldSDFWriterUtility import write_sdf
 from blenderproc.python.writer.WriterUtility import write_hdf5
