### 2022-10-27 945f2c6
remove sdf writer module
Signed-off-by: Mohamad <mohamad@openrobotics.org>
