### 2022-10-28 #4783 @themasterlink
> @themasterlink could you please help suggest reviewers too? point_up thanks

@Kevin-Mattheus-Moerman Sorry, how should I suggest a reviewer? I am the person who wrote the paper. I can not select the person who reviews it. That would make the whole review process worthless.

If we ignore that, I would pick @natevm, I don't know him personally, and he is part of the team behind NVISII, which is similar to BlenderProc, so he should be able to review the paper.

But, still, it seems very odd to me that I should pick my reviewer :D
