### 2022-10-28 #4783 @Kevin-Mattheus-Moerman
@themasterlink apologies for the confusion. It is not that you "get to pick/assign" reviewers, that remains up to the editor, i.e. me, however it is common for journals to accept reviewer suggestions from the authors. I hope that clarifies it. I also asked not to use the `@` symbol to avoid tagging them, so if you have any other suggestions please just list their github name and I can explore whether to invite them here. Thanks.
