### 2022-10-28 #4783 @Kevin-Mattheus-Moerman
@BradyAJohnston @xqms @SelvamArul @aadi-mishra @tduboudi @xheon @natevm would you be interested in reviewing this submission for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/)? The work in question is titled: **"BlenderProc2: A Procedural Pipeline for Photorealistic Rendering"**, and has applications to deep learning methods.

JOSS reviews happen on GitHub and focus on reviewing the [software](https://github.com/DLR-RM/BlenderProc) as well as a [short paper](https://github.com/openjournals/joss-papers/blob/joss.04783/joss.04783/10.21105.joss.04783.pdf).

For more information on our review process see our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

You can let me know in this issue if you are interested in reviewing this work.
