### 2022-11-01 #4783 @natevm
I’d be interested !-NateOn Nov 1, 2022, at 2:03 AM, Kevin Mattheus Moerman ***@***.***> wrote:﻿
@xqms @SelvamArul @aadi-mishra @tduboudi @xheon @natevm would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)? The work in question is titled: "BlenderProc2: A Procedural Pipeline for Photorealistic Rendering", and has applications to deep learning methods.
JOSS reviews happen on GitHub and focus on reviewing the software as well as a short paper.
For more information on our review process see our review guidelines.
You can let me know in this issue if you are interested in reviewing this work.
