### 2022-11-02 #4901 @editorialbot
[REVIEW]: BlenderProc2
**Submitting author:** @themasterlink (<a href="http://orcid.org/0000-0002-1557-2234">Maximilian Denninger</a>)
**Repository:** https://github.com/DLR-RM/BlenderProc
**Branch with paper.md** (empty if default branch):
**Version:** v2.4.1

@nicoguaro & @BradyAJohnston & @natevm & @SelvamArul, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
