### 2022-11-11 c19f3eb
fix(utility): Removes obsolete code
