### 2022-11-11 fc6ea24
chore(utility): Adds missing docu
