### 2022-11-11 c87cab8
chore(utility): Renames function name and correct exception class
