### 2022-11-11 dfca734
chore(utility): Simplifies get_file_descriptor()
