### 2022-11-11 c061208
feat(renderer): Adds progress bar to visualize render progress
