+++ blenderproc/python/renderer/RendererUtility.py
@@ -1,7 +1,9 @@
 """Provides functionality to render a color, normal, depth and distance image."""
 
+from contextlib import contextmanager
 import os
-from typing import Union, Dict, List, Set, Optional, Any
+import threading
+from typing import IO, Union, Dict, List, Set, Optional, Any
 import math
 import sys
 import platform
@@ -10,6 +12,9 @@ import time
 import mathutils
 import bpy
 import numpy as np
+from rich.console import Console
+from rich.progress import Progress, TextColumn, BarColumn, TimeRemainingColumn 
+from io import StringIO
 
 from blenderproc.python.camera import CameraUtility
 from blenderproc.python.modules.main.GlobalStorage import GlobalStorage
@@ -524,6 +529,93 @@ def map_file_format_to_file_ending(file_format: str) -> str:
     raise RuntimeError(f"Unknown Image Type {file_format}")
 
 
+def _progress_bar_thread(pipe_out: int, stdout: IO, total_frames: int, num_samples: int):
+    """ The thread rendering the progress bar
+
+    :param pipe_out: The pipe output delivering blenders debug messages.
+    :param stdout: The stdout to which the progress bar should be written.
+    :param total_frames: The number of frames that should be rendered.
+    :param num_samples: The number of samples used to render each frame.
+    """
+    # Define columns for progress bar
+    columns = [
+        TextColumn("[progress.description]{task.description}"),
+        BarColumn(),
+        TimeRemainingColumn(),
+        TextColumn("[progress.description]{task.fields[status]}"),
+    ]
+    # Initializes progress bar using given stdout
+    with Progress(*columns, console=Console(file=stdout), transient=True) as progress:
+        complete_task = progress.add_task("[green]Total", total=total_frames, status="")
+        frame_task = progress.add_task("[yellow]Current frame", total=num_samples, status="")
+
+        # Continuously read blenders debug messages
+        current_line = ""
+        while True:
+            # Read the next character
+            char = os.read(pipe_out, 1).decode()
+            
+            # If its the ending character, stop
+            if not char or "\b" == char:
+                break
+            # If the current line has ended
+            if char == "\n":                
+                # Check if its a line we can use (starts with "Fra:")
+                if current_line.startswith("Fra:"):
+                    # Extract current frame number and set to progress bar
+                    frame_number = int(current_line.split()[0][len("Fra:"):])
+                    progress.update(complete_task, completed=frame_number)
+                    progress.update(complete_task, status=f"Rendering frame {frame_number + 1} of {total_frames}")
+
+                    # Split line into columns
+                    status_columns = [col.strip() for col in current_line.split("|")]
+                    if "Scene, ViewLayer" in status_columns:
+                        # If we are currently at "Scene, ViewLayer", use everything afterwards
+                        status = " | ".join(status_columns[status_columns.index("Scene, ViewLayer") + 1:])
+                        # If we are currently rendering, update the progress 
+                        if status.startswith("Sample"):
+                            progress.update(frame_task, completed=int(status[len("Sample"):].split("/")[0]))
+                    elif "Compositing" in status_columns:
+                        # If we are at "Compositing", use everything afterwards including "Compositing"
+                        status = " | ".join(status_columns[status_columns.index("Compositing"):])
+                        # Set render progress to complete
+                        progress.update(frame_task, completed=num_samples)
+                    else:
+                        # In every other case, use last column
+                        status = status_columns[-1]
+                    # Set status to progress bar
+                    progress.update(frame_task, status=status)
+                # Start with next line
[…]
