### 2022-11-14 4781ba7
Fix(PostProcessing): trim redundant channel preserves input type
