### 2022-11-14 20d8f97
Feat(CocoWriter): raise error if stereo images are given
