### 2022-11-14 21bd61f
Feat(stereo_matching): Add segmentation output as test
