### 2022-11-14 a22c7f2
Fix(visHdf5): Fix stereo label visualization
