### 2022-11-24 d079293
chore(github): Puts log files into artifacts
