+++ .github/workflows/blenderprochelper.yml
@@ -11,3 +11,8 @@ jobs:
     - uses: actions/checkout@v2
     - name: Run BlenderProcHelper
       run: python /home/$USER/workspace/BlenderProcHelper/github_action.py
+    - name: Archive logs
+      uses: actions/upload-artifact@v3
+      with:
+        name: logs
+        path: artifacts
