### 2022-11-24 8e30bf8
fix(coco): Fixes cocowriter usage in yaml files
