### 2022-11-25 42dd86f
feat(writer): add option to indent COCO annotations
New param in write_coco_annotations() in CocoWriterUtility.py
to enable indents when writing annotations

Fix typo in CONTRIBUTING.md

closes #764
