### 2022-11-28 69bd9d6
fix(bop): Fixes bop writers camera.json when using yaml configs
