+++ blenderproc/python/writer/CocoWriterUtility.py
@@ -62,8 +62,8 @@ def write_coco_annotations(output_dir: str, instance_segmaps: Optional[List[np.n
         instance_attribute_maps = []
 
     if len(colors) > 0 and len(colors[0].shape) == 4:
-        raise ValueError(f"BlenderProc currently does not support writing coco annotations for stereo images. "
-                         f"However, you can enter left and right images / segmaps separately.")
+        raise ValueError("BlenderProc currently does not support writing coco annotations for stereo images. "
+                         "However, you can enter left and right images / segmaps separately.")
 
     # Create output directory
     os.makedirs(os.path.join(output_dir, 'images'), exist_ok=True)
