### 2022-12-02 407d784
fix(types): fix return type of get_all_cps()
- change the return type of get_all_cps() to be dict
- now returns dict instead of IDPropertyGroupViewItems
- Update documentation to match
