+++ examples/advanced/coco_annotations/main.py
@@ -15,7 +15,7 @@ objs = bproc.loader.load_blend(args.scene)
 
 # Set some category ids for loaded objects
 for j, obj in enumerate(objs):
-    obj.set_cp("category_id", j+1)
+    obj.set_cp("category_id", j + 1)
 
 # define a light and set its location and energy level
 light = bproc.types.Light()
