### 2022-12-06 5ac0b38
fix(struct): Fixes pylint complaints
