### 2022-12-13 231f41c
feat(logo): adds code for rendering blenderproc logo
