### 2022-12-13 d8d1e92
feat(logo): adds code for rendering blenderproc logo
