+++ blenderproc/python/renderer/NOCSRendererUtility.py
@@ -1,0 +1,109 @@
+"""Provides functionality to render a Normalized Object Coordinate Space (NOCS) image."""
+
+from typing import Optional, Dict, List
+
+import bpy
+import numpy as np
+
+from blenderproc.python.material import MaterialLoaderUtility
+from blenderproc.python.renderer import RendererUtility
+from blenderproc.python.types.MaterialUtility import Material
+from blenderproc.python.utility.BlenderUtility import get_all_blender_mesh_objects
+from blenderproc.python.utility.Utility import Utility
+
+
+def set_world_background(color: List[float], strength: float = 1):
+    """ Sets the color of blenders world background
+
+    :param color: A three-dimensional list specifying the new color in floats.
+    :param strength: The strength of the emitted background light.
+    """
+    world = bpy.context.scene.world
+    world.use_nodes = True
+    nodes = world.node_tree.nodes
+    links = world.node_tree.links
+
+    # Unlink any incoming link that would overwrite the default value
+    if len(nodes.get("Background").inputs['Color'].links) > 0:
+        links.remove(nodes.get("Background").inputs['Color'].links[0])
+
+    nodes.get("Background").inputs['Strength'].default_value = strength
+    nodes.get("Background").inputs['Color'].default_value = color + [1]
+
+def render_nocs(output_dir: Optional[str] = None, file_prefix: str = "nocs_", output_key: str = "nocs",
+                return_data: bool = True, verbose: bool = False) -> Dict[str, List[np.ndarray]]:
+    """ Renders the Normalized Object Coordinate Space (NOCS).
+
+    Colors each object based on its local coordinates.
+    The coordinates [-1, 1] are mapped into the [0, 1] colorspace.
+    It is therefore, recommended that all local vertex coordinates are in range [-1, 1].
+    The world background is rendered transparent.
+
+    :param output_dir: The directory to write images to. If None is given, the temp dir is used.
+    :param file_prefix: The prefix to use for writing the images.
+    :param output_key: The key to use for registering the output.
+    :param return_data: Whether to load and return generated data.
+    :param verbose: If True, more details about the rendering process are printed.
+    :return: A dict containing one entry "nocs" which points to the list of rendered frames.
+    """
+    if output_dir is None:
+        output_dir = Utility.get_temporary_directory()
+
+    with Utility.UndoAfterExecution():
+        nocs_material = _NOCSRendererUtility.create_nocs_material()
+
+        # Set the NOCS material to all objects
+        for obj in get_all_blender_mesh_objects():
+            if len(obj.material_slots) > 0:
+                for i in range(len(obj.material_slots)):
+                    obj.data.materials[i] = nocs_material.blender_obj
+            else:
+                obj.data.materials.append(nocs_material.blender_obj)
+
+        # Make sure the background is black
+        set_world_background([0, 0, 0])
+
+        # Set all fast rendering parameters with only one ray per pixel
+        RendererUtility._render_init()
+        # the amount of samples must be one and there can not be any noise threshold
+        RendererUtility.set_samples(1)
+        RendererUtility.set_denoiser(None)
+        RendererUtility.set_light_bounces(1, 0, 0, 1, 0, 8, 0)
+        bpy.context.scene.cycles.filter_width = 0.0
+
+        # Use exr as output format, as it uses a linear colorspace and uses float16
+        RendererUtility.set_output_format("OPEN_EXR", 16, enable_transparency=True)
+        # Render and ret
+        return RendererUtility.render(output_dir, file_prefix, output_key, load_keys={output_key},
[…]
