### 2022-12-13 6d23daa
fix(tests): Fixes precision of assertAlmostEqual
