### 2022-12-13 829accd
chore(tests): Adds docu about how to run unit tests
