+++ README.md
@@ -235,17 +235,6 @@ If you use BlenderProc in a research project, please cite as follows:
 }
 ```
 
-## Unit tests
-
-The unit tests located in the `tests/` folder can be run via the following command:
-
-```
-BP_CC_MATERIALS_PATH=<cc_matrials_path> BP_HAVEN_PATH=<haven_path>  python3 cli.py run tests/run_all.py
-```
-
-where `<cc_matrials_path>` and `<haven_path>` are paths to the respective datasets.
-You can download them using blenderprocs [download command](docs/tutorials/loader.md).
-
 ---
 
 <div align="center">
