### 2022-12-13 48f07f9
Updates table1 in paper
