### 2022-12-13 d3f9fe7
chore(paper): Adds gui-based debugger row to table 1
