### 2022-12-13 #4901 @cornerfarmer
Hey @nicoguaro,

thank you for your comments!

> I can't find a mention to how to test the software in the documentation.

You are right I added a respective [readme file](https://github.com/DLR-RM/BlenderProc/tree/main/tests)

>  It would be nice to have scripts to generate Figure 1 in the paper and the first image on the README of the repo. There are plenty of examples, but they are the very first impressions for the software, so...

These figures are actually based on an already existing example: https://github.com/DLR-RM/BlenderProc/tree/main/examples/datasets/front_3d_with_improved_mat
It shows how to load scenes from the 3D-Front dataset, how to sample camera poses in them and then how to render all kinds of modalities

I also slightly updated table 1 in the paper
