### 2022-12-21 4d5a5c0
Get env:USER in a more reliable way
In some environment like Docker, $USER might not be set
https://stackoverflow.com/questions/54411218/docker-why-isnt-user-environment-variable-set
`getpass.getuser` checks the environment variables LOGNAME, USER, LNAME and USERNAME, in order, and returns the value of the first one which is set to a non-empty string. If none are set, the login name from the password database is returned on systems which support the pwd module, otherwise, an exception is raised.
