### 2023-01-16 2536f6c
Chore(README): fool proof installation instructions
