### 2023-01-18 1537907
feat(hdri): Allows setting brightness of hdri backgrounds
