### 2023-01-19 d1e8d86
fix(loader): fix spelling error in cocowriter
Corrects a spelling error in CocoWriterUtility.py. Also fixes a likely copy-paste error in CONTRIBUTING.md

closes #785
