### 2023-01-25 988137f
Add texture loading to .ply objects.
