+++ blenderproc/python/loader/ObjectLoader.py
@@ -1,16 +1,26 @@
 """Provides `load_obj`, which allows to load different 3D object files. """
 
 import os
+import re
 from typing import List, Optional, Dict
 
 import bpy
 
 from blenderproc.python.types.MeshObjectUtility import MeshObject, convert_to_meshes
+from blenderproc.python.types.MaterialUtility import Material
+from blenderproc.python.utility.Utility import Utility
+from blenderproc.python.material.MaterialLoaderUtility import (
+    create_material_from_texture,
+)
 
 
-def load_obj(filepath: str, cached_objects: Optional[Dict[str, List[MeshObject]]] = None,
-             use_legacy_obj_import: bool = False, **kwargs) -> List[MeshObject]:
-    """ Import all objects for the given file and returns the loaded objects
+def load_obj(
+    filepath: str,
+    cached_objects: Optional[Dict[str, List[MeshObject]]] = None,
+    use_legacy_obj_import: bool = False,
+    **kwargs,
+) -> List[MeshObject]:
+    """Import all objects for the given file and returns the loaded objects
 
     In .obj files a list of objects can be saved in.
     In .ply files only one object can be saved so the list has always at most one element
@@ -23,47 +33,102 @@ def load_obj(filepath: str, cached_objects: Optional[Dict[str, List[MeshObject]]
     :param kwargs: all other params are handed directly to the bpy loading fct. check the corresponding documentation
     :return: The list of loaded mesh objects.
     """
-    if os.path.exists(filepath):
-        if cached_objects is not None and isinstance(cached_objects, dict):
-            if filepath in cached_objects.keys():
-                created_obj = []
-                for obj in cached_objects[filepath]:
-                    # duplicate the object
-                    created_obj.append(obj.duplicate())
-                return created_obj
-            loaded_objects = load_obj(filepath, cached_objects=None, **kwargs)
-            cached_objects[filepath] = loaded_objects
-            return loaded_objects
-        # save all selected objects
-        previously_selected_objects = bpy.context.selected_objects
-        if filepath.endswith('.obj'):
-            # load an .obj file:
-            if use_legacy_obj_import:
-                bpy.ops.import_scene.obj(filepath=filepath, **kwargs)
-            else:
-                bpy.ops.wm.obj_import(filepath=filepath, **kwargs)
-        elif filepath.endswith('.ply'):
+    if not os.path.exists(filepath):
+        raise FileNotFoundError(f"The given filepath does not exist: {filepath}")
+
+    if cached_objects is not None and isinstance(cached_objects, dict):
+        if filepath in cached_objects.keys():
+            created_obj = []
+            for obj in cached_objects[filepath]:
+                # duplicate the object
+                created_obj.append(obj.duplicate())
+            return created_obj
+        loaded_objects = load_obj(filepath, cached_objects=None, **kwargs)
+        cached_objects[filepath] = loaded_objects
+        return loaded_objects
+    # save all selected objects
+    previously_selected_objects = bpy.context.selected_objects
+    if filepath.endswith(".obj"):
+        # load an .obj file:
+        if use_legacy_obj_import:
+            bpy.ops.import_scene.obj(filepath=filepath, **kwargs)
+        else:
+            bpy.ops.wm.obj_import(filepath=filepath, **kwargs)
+    elif filepath.endswith(".ply"):
+        PLY_TEXTURE_FILE_COMMENT = "comment TextureFile "
+        model_name = os.path.basename(filepath)
+
+        # Read file
+        with open(filepath, "r", encoding="latin-1") as file:
+            ply_file_content = file.read()
+
+        # Check if texture file is given
+        if PLY_TEXTURE_FILE_COMMENT in ply_file_content:
[…]
