### 2023-01-26 c6e1ce6
Use ObjectLoader.load_obj inside BopLoader
