+++ blenderproc/python/loader/BopLoader.py
@@ -3,24 +3,33 @@
 import os
 from random import choice
 from typing import List, Optional, Tuple
-
 import bpy
 import numpy as np
 from mathutils import Matrix, Vector
 
 from blenderproc.python.utility.SetupUtility import SetupUtility
+from blenderproc.python.loader.ObjectLoader import load_obj
 from blenderproc.python.camera import CameraUtility
 from blenderproc.python.types.MeshObjectUtility import MeshObject
 from blenderproc.python.utility.Utility import Utility
-from blenderproc.python.utility.MathUtility import change_source_coordinate_frame_of_transformation_matrix
+from blenderproc.python.utility.MathUtility import (
+    change_source_coordinate_frame_of_transformation_matrix,
+)
 from blenderproc.python.types.MaterialUtility import Material
 
 
-def load_bop_objs(bop_dataset_path: str, model_type: str = "", obj_ids: Optional[List[int]] = None,
-                  sample_objects: bool = False, num_of_objs_to_sample: Optional[int] = None,
-                  obj_instances_limit: int = -1, mm2m: bool = False, move_origin_to_x_y_plane: bool = False,
-                  temp_dir: Optional[str] = None) -> List[MeshObject]:
-    """ Loads all or a subset of 3D models of any BOP dataset
+def load_bop_objs(
+    bop_dataset_path: str,
+    model_type: str = "",
+    obj_ids: Optional[List[int]] = None,
+    sample_objects: bool = False,
+    num_of_objs_to_sample: Optional[int] = None,
+    obj_instances_limit: int = -1,
+    mm2m: bool = False,
+    move_origin_to_x_y_plane: bool = False,
+    temp_dir: Optional[str] = None,
+) -> List[MeshObject]:
+    """Loads all or a subset of 3D models of any BOP dataset
 
     :param bop_dataset_path: Full path to a specific bop dataset e.g. /home/user/bop/tless.
     :param model_type: Optionally, specify type of BOP model. Available: [reconst, cad or eval].
@@ -42,9 +51,12 @@ def load_bop_objs(bop_dataset_path: str, model_type: str = "", obj_ids: Optional
     # This import is done inside to avoid having the requirement that BlenderProc depends on the bop_toolkit
     # pylint: disable=import-outside-toplevel
     from bop_toolkit_lib import dataset_params
+
     # pylint: enable=import-outside-toplevel
 
-    model_p = dataset_params.get_model_params(bop_path, bop_dataset_name, model_type=model_type if model_type else None)
+    model_p = dataset_params.get_model_params(
+        bop_path, bop_dataset_name, model_type=model_type if model_type else None
+    )
 
     if temp_dir is None:
         temp_dir = Utility.get_temporary_directory()
@@ -55,35 +67,56 @@ def load_bop_objs(bop_dataset_path: str, model_type: str = "", obj_ids: Optional
         obj_ids = []
     allow_duplication = obj_ids or sample_objects
 
-    obj_ids = obj_ids if obj_ids else model_p['obj_ids']
+    obj_ids = obj_ids if obj_ids else model_p["obj_ids"]
 
     loaded_objects = []
     # if sampling is enabled
     if sample_objects:
         loaded_ids = {}
         loaded_amount = 0
-        if obj_instances_limit != -1 and len(obj_ids) * obj_instances_limit < num_of_objs_to_sample:
-            raise RuntimeError(f"{bop_dataset_path}'s contains {len(obj_ids)} objects, {num_of_objs_to_sample} object "
-                               f"where requested to sample with an instances limit of {obj_instances_limit}. Raise "
-                               f"the limit amount or decrease the requested amount of objects.")
+        if (
+            obj_instances_limit != -1
[…]
