### 2023-01-27 ae6e375
feat(#784): integrate bop_toolkit scripts into the bop writer
