+++ blenderproc/python/writer/BopWriterUtility.py
@@ -6,6 +6,7 @@ import glob
 from typing import List, Optional
 import shutil
 import warnings
+import datetime
 
 import numpy as np
 import png
@@ -18,13 +19,15 @@ from blenderproc.python.utility.Utility import Utility, resolve_path
 from blenderproc.python.postprocessing.PostProcessingUtility import dist2depth
 from blenderproc.python.writer.WriterUtility import _WriterUtility
 from blenderproc.python.types.LinkUtility import Link
+from blenderproc.python.loader.BopLoader import _BopLoader
 
 
 def write_bop(output_dir: str, target_objects: Optional[List[MeshObject]] = None,
               depths: Optional[List[np.ndarray]] = None, colors: Optional[List[np.ndarray]] = None,
               color_file_format: str = "PNG", dataset: str = "", append_to_existing_output: bool = True,
               depth_scale: float = 1.0, jpg_quality: int = 95, save_world2cam: bool = True,
-              ignore_dist_thres: float = 100., m2mm: bool = True, frames_per_chunk: int = 1000):
+              ignore_dist_thres: float = 100., m2mm: bool = True, frames_per_chunk: int = 1000,
+              calc_mask_info_coco: bool = False, delta: int = 15):
     """Write the BOP data
 
     :param output_dir: Path to the output directory.
@@ -46,6 +49,8 @@ def write_bop(output_dir: str, target_objects: Optional[List[MeshObject]] = None
     :param m2mm: Original bop annotations and models are in mm. If true, we convert the gt annotations to mm here. This
                  is needed if BopLoader option mm2m is used.
     :param frames_per_chunk: Number of frames saved in each chunk (called scene in BOP)
+    :param calc_mask_info_coco: Whether to calculate gt masks, gt info and gt coco annotations.
+    :param delta: Tolerance used for estimation of the visibility masks.
     """
     if depths is None:
         depths = []
@@ -82,12 +87,54 @@ def write_bop(output_dir: str, target_objects: Optional[List[MeshObject]] = None
                            f"Either remove the dataset parameter or assign custom property 'bop_dataset_name'"
                            f" to selected objects")
 
+    if calc_mask_info_coco:
+        # It might be that a chunk dir already exists where the writer appends frames.
+        # If one (or multiple) more chunk dirs are created to save the rendered frames to,
+        # mask/info/coco annotations need to be calculated for all of them
+        chunk_dirs = sorted(glob.glob(os.path.join(chunks_dir, '*')))
+        chunk_dirs = [d for d in chunk_dirs if os.path.isdir(d)]
+        starting_chunk_dir = sorted(chunk_dirs)[-1] if chunk_dirs else None
+
     # Save the data.
     _BopWriterUtility.write_camera(camera_path, depth_scale=depth_scale)
     _BopWriterUtility.write_frames(chunks_dir, dataset_objects=dataset_objects, depths=depths, colors=colors,
                                    color_file_format=color_file_format, frames_per_chunk=frames_per_chunk,
                                    m2mm=m2mm, ignore_dist_thres=ignore_dist_thres, save_world2cam=save_world2cam,
                                    depth_scale=depth_scale, jpg_quality=jpg_quality)
+
+    if calc_mask_info_coco:
+        # Set up the bop toolkit
+        _BopLoader.setup_bop_toolkit(dataset_dir)
+
+        # This import is done inside to avoid having the requirement that BlenderProc depends on the bop_toolkit
+        # pylint: disable=import-outside-toplevel
+        from bop_toolkit_lib import misc, renderer
+        # pylint: enable=import-outside-toplevel
+
+        # Initialize a renderer.
[…]
