### 2023-01-27 c87db65
Remove depreciated function parameters from load_mesh
