### 2023-01-27 2bbe90b
Change material name in BopLoader to be backward compatible with .yaml based examples
