### 2023-01-27 3b64070
Change material creation to use blenderproc material creator
