### 2023-01-27 e2e2519
Remove previously_selected_objects
Removed previously_selected_objects to see if it has an impact or does not have to be used.
