### 2023-01-27 764bb4c
Remove unused functions from _BopLoader
