+++ blenderproc/python/loader/BopLoader.py
@@ -357,61 +357,3 @@ class _BopLoader:
         cur_obj.set_cp("bop_dataset_name", bop_dataset_name)
 
         return cur_obj
-
-    @staticmethod
-    def load_materials(cur_obj: bpy.types.Object, bop_dataset_name: str) -> Material:
-        """ Loads / defines materials, e.g. vertex colors.
-
-        :param cur_obj: The object to use.
-        :param bop_dataset_name: The name of the used bop dataset.
-        :return: Material with vertex color.
-        """
-
-        mat = cur_obj.data.materials.get("Material")
-
-        if mat is None:
-            # create material
-            mat = bpy.data.materials.new(name="bop_" + bop_dataset_name + "_vertex_col_material")
-
-        mat.use_nodes = True
-
-        if cur_obj.data.materials:
-            # assign to 1st material slot
-            cur_obj.data.materials[0] = mat
-        else:
-            # no slots
-            cur_obj.data.materials.append(mat)
-
-        return Material(mat)
-
-    @staticmethod
-    def load_texture(cur_obj: bpy.types.Object, texture_file_path: str, bop_dataset_name: str):
-        """
-        Load the textures for the ycbv objects, only those contain texture information
-
-        :param cur_obj: The object to use.
-        :param texture_file_path: path to the texture file (most likely ".png")
-        :param bop_dataset_name: The name of the used bop dataset.
-        """
-        mat = bpy.data.materials.new(name="bop_" + bop_dataset_name + "_texture_material")
-
-        mat.use_nodes = True
-
-        nodes = mat.node_tree.nodes
-        links = mat.node_tree.links
-
-        color_image = nodes.new('ShaderNodeTexImage')
-        if not os.path.exists(texture_file_path):
-            raise FileNotFoundError(f"The texture path for the ycbv object could not be loaded "
-                                    f"from the file: {texture_file_path}")
-        color_image.image = bpy.data.images.load(texture_file_path, check_existing=True)
-
-        principled = Utility.get_the_one_node_with_type(nodes, "BsdfPrincipled")
-        links.new(color_image.outputs["Color"], principled.inputs["Base Color"])
-
-        if cur_obj.data.materials:
-            # assign to 1st material slot
-            cur_obj.data.materials[0] = mat
-        else:
-            # no slots
-            cur_obj.data.materials.append(mat)
