### 2023-01-27 00e5235
Remove allow_duplication to test material dependency on mesh
