### 2023-01-27 a17dd7e
Update blenderproc/python/loader/ObjectLoader.py
