+++ blenderproc/python/loader/ObjectLoader.py
@@ -57,7 +57,7 @@ def load_obj(filepath: str, cached_objects: Optional[Dict[str, List[MeshObject]]
             # add a default material to stl file
             mat = bpy.data.materials.new(name="stl_material")
             mat.use_nodes = True
-            selected_objects = [obj for obj in bpy.context.selected_objects if]
+            selected_objects = [obj for obj in bpy.context.selected_objects]
             for obj in selected_objects:
                 obj.data.materials.append(mat)
         return convert_to_meshes([obj for obj in bpy.context.selected_objects])
