### 2023-01-30 2e6dfc6
refactor(#784): calculate mask_info_coco per default
