### 2023-01-30 5c4e0f1
refactor(#784): use setup_bop_toolkit to install the bop_toolkit
