### 2023-01-30 c2976a2
Update blenderproc/python/loader/BopLoader.py
Co-authored-by: Dominik Winkelbauer <dominik.winkelbauer@outlook.de>
