### 2023-01-30 f238ada
fix(BopLoader): fix materials being overwritten by mesh duplication
Previous commits lead to wrong materials for duplicated objects,
therefore we integrated the cached_objects functionality from load_obj
instead and duplicate the materials from the duplicated objects. See
Issue 590 for a related discussion.
