### 2023-01-30 60b833f
format(loader): Format loader files to resolve pylint warning
