### 2023-01-30 6273c3b
fix(bop_loader_module): fix unexpected keyword error temp_dir
The old config based examples set the temp_dir variable although it's not used anymore. Removed the usage of the function inside BopLoaderModule, should be removed from the config based examples in the future.
