### 2023-01-30 ca24cfe
fix(bop_loader): Removes unused positional argument
After removing 'has_external_texture' the function call had to be adopted. It is now automatically determined if the object has a texture or not.
