### 2023-01-30 ea5b60d
fix(bop_loader_module): Remove unused argument
Another one...
