### 2023-01-30 fd21ff2
refactor(#784): use SetupUtility directly to set up the bop_toolkit
