### 2023-01-30 349b693
refactor(#784): use SetupUtility directly to set up the bop_toolkit
