### 2023-01-30 3ddc273
refactor(#784): improve singleton instance handling
