### 2023-01-30 890684e
style(bop_loader): Fix pylint line-too-long
