### 2023-01-30 57729df
style(bop_loader): Fix pylint consider-iterating-dictionary
