### 2023-02-01 3aca723
refactor(#784): only calculate mask info coco for newly saved frames
