+++ blenderproc/python/writer/BopWriterUtility.py
@@ -93,7 +93,21 @@ def write_bop(output_dir: str, target_objects: Optional[List[MeshObject]] = None
         # mask/info/coco annotations need to be calculated for all of them
         chunk_dirs = sorted(glob.glob(os.path.join(chunks_dir, '*')))
         chunk_dirs = [d for d in chunk_dirs if os.path.isdir(d)]
-        starting_chunk_dir = sorted(chunk_dirs)[-1] if chunk_dirs else None
+        last_chunk_dir = sorted(chunk_dirs)[-1] if chunk_dirs else None
+
+        starting_chunk_id = 0
+        starting_frame_id = 0
+        if last_chunk_dir:
+            last_chunk_gt_fpath = os.path.join(last_chunk_dir, 'scene_gt.json')
+            chunk_gt = _BopWriterUtility.load_json(last_chunk_gt_fpath, keys_to_int=True)
+
+            # Current chunk and frame ID's.
+            starting_chunk_id = int(os.path.basename(last_chunk_dir))
+            starting_frame_id = int(sorted(chunk_gt.keys())[-1]) + 1
+
+            if starting_frame_id % frames_per_chunk == 0:
+                starting_chunk_id += 1
+                starting_frame_id = 0
 
     # Save the data.
     _BopWriterUtility.write_camera(camera_path, depth_scale=depth_scale)
@@ -107,12 +121,12 @@ def write_bop(output_dir: str, target_objects: Optional[List[MeshObject]] = None
         SetupUtility.setup_pip(["git+https://github.com/thodan/bop_toolkit", "vispy>=0.6.5",
                                 "PyOpenGL==3.1.0"])
 
-        _BopWriterUtility.calc_gt_masks(chunks_dir=chunks_dir, starting_chunk_dir=starting_chunk_dir,
-                                        dataset_objects=dataset_objects, delta=delta)
-        _BopWriterUtility.calc_gt_info(chunks_dir=chunks_dir, starting_chunk_dir=starting_chunk_dir,
-                                       dataset_objects=dataset_objects, delta=delta)
-        _BopWriterUtility.calc_gt_coco(chunks_dir=chunks_dir, starting_chunk_dir=starting_chunk_dir,
-                                       dataset_objects=dataset_objects)
+        _BopWriterUtility.calc_gt_masks(chunks_dir=chunks_dir, starting_chunk_id=starting_chunk_id,
+                                        starting_frame_id=starting_frame_id, dataset=dataset, delta=delta)
+        _BopWriterUtility.calc_gt_info(chunks_dir=chunks_dir, starting_chunk_id=starting_chunk_id,
+                                       starting_frame_id=starting_frame_id, dataset=dataset, delta=delta)
+        _BopWriterUtility.calc_gt_coco(chunks_dir=chunks_dir, dataset_objects=dataset_objects,
+                                       starting_chunk_id=starting_chunk_id, starting_frame_id=starting_frame_id)
 
 class _BopWriterUtility:
     """ Saves the synthesized dataset in the BOP format. The dataset is split
@@ -451,14 +465,16 @@ class _BopWriterUtility:
                 curr_frame_id += 1
 
     @staticmethod
-    def calc_gt_masks(chunks_dir: str, starting_chunk_dir: Optional[str], dataset_objects: List[MeshObject],
-                      delta: int = 15):
+    def calc_gt_masks(chunks_dir: str, starting_chunk_id: Optional[str], starting_frame_id: Optional[int],
+                      dataset: str = "", delta: int = 15):
         """ Calculates the ground truth masks.
         From the BOP toolkit (https://github.com/thodan/bop_toolkit).
 
         :param chunks_dir: Path to the output directory of the current chunk.
[…]
