### 2023-02-01 ac3c02f
fix(#784): calculate gt info coco for all dataset objects
