### 2023-02-01 92dfe7d
fix(BopWriterUtility): fixed wrong getter for "category_id" inside get_frame_gt
