### 2023-02-01 880d1b9
Delete debug.sh
