### 2023-02-01 9588027
Delete debug_config.sh
