### 2023-02-02 4ca8b5f
refactor(#784): move shared code outside
