+++ blenderproc/python/writer/BopWriterUtility.py
@@ -121,12 +121,33 @@ def write_bop(output_dir: str, target_objects: Optional[List[MeshObject]] = None
         SetupUtility.setup_pip(["git+https://github.com/thodan/bop_toolkit", "vispy>=0.6.5",
                                 "PyOpenGL==3.1.0"])
 
-        _BopWriterUtility.calc_gt_masks(chunks_dir=chunks_dir, starting_chunk_id=starting_chunk_id,
-                                        starting_frame_id=starting_frame_id, dataset=dataset, delta=delta)
-        _BopWriterUtility.calc_gt_info(chunks_dir=chunks_dir, starting_chunk_id=starting_chunk_id,
-                                       starting_frame_id=starting_frame_id, dataset=dataset, delta=delta)
-        _BopWriterUtility.calc_gt_coco(chunks_dir=chunks_dir, dataset_objects=dataset_objects,
-                                       starting_chunk_id=starting_chunk_id, starting_frame_id=starting_frame_id)
+        # determine which objects to add to the vsipy renderer
+        # for numpy>=1.20, np.float is deprecated: https://numpy.org/doc/stable/release/1.20.0-notes.html#deprecations
+        np.float = float
+        dataset_objects = []
+        if dataset != "":
+            for obj in get_all_mesh_objects():
+                if "bop_dataset_name" in obj.blender_obj:
+                    # opposite to above, we write annotations also for hidden objects
+                    if obj.blender_obj["bop_dataset_name"] == dataset:
+                        dataset_objects.append(obj)
+        else:
+            # if no dataset is defined, write annotations for all objects
+            dataset_objects = get_all_mesh_objects()
+
+        # Determine for which directories mask_info_coco has to be calculated
+        chunk_dirs = sorted(glob.glob(os.path.join(chunks_dir, '*')))
+        chunk_dirs = [d for d in chunk_dirs if os.path.isdir(d)]
+        chunk_dir_ids = [d.split('/')[-1] for d in chunk_dirs]
+        chunk_dirs = chunk_dirs[chunk_dir_ids.index(f"{starting_chunk_id:06d}"):]
+
+        _BopWriterUtility.calc_gt_masks(chunk_dirs=chunk_dirs, starting_frame_id=starting_frame_id,
+                                        dataset_objects=dataset_objects, delta=delta)
+        _BopWriterUtility.calc_gt_info(chunk_dirs=chunk_dirs, starting_frame_id=starting_frame_id,
+                                       dataset_objects=dataset_objects, delta=delta)
+        _BopWriterUtility.calc_gt_coco(chunk_dirs=chunk_dirs, dataset_objects=dataset_objects,
+                                       starting_frame_id=starting_frame_id)
+
 
 class _BopWriterUtility:
     """ Saves the synthesized dataset in the BOP format. The dataset is split
@@ -465,16 +486,14 @@ class _BopWriterUtility:
                 curr_frame_id += 1
 
     @staticmethod
-    def calc_gt_masks(chunks_dir: str, starting_chunk_id: Optional[str], starting_frame_id: Optional[int],
-                      dataset: str = "", delta: int = 15):
+    def calc_gt_masks(chunk_dirs: List[str], dataset_objects: List[MeshObject], starting_frame_id: int = 0,
+                      delta: int = 15):
         """ Calculates the ground truth masks.
         From the BOP toolkit (https://github.com/thodan/bop_toolkit).
 
-        :param chunks_dir: Path to the output directory of the current chunk.
-        :param starting_chunk_id: The first chunk_dir the writer has written to during this run.
+        :param chunk_dirs: List of directories to calculate the gt masks for.
+        :param dataset_objects: Save annotations for these objects.
[…]
