### 2023-02-02 #4901 @natevm
I was able to check all these boxes as well. One small comment, in Table 1, the authors claim NVISII's renderer is not open source. That is not true though, the code for NVISII is available here: https://github.com/owl-project/NVISII

Perhaps we could get that corrected before acceptance?
