### 2023-02-03 f431737
fix(#810): Remove vispy reinitialization
Wait until bop_toolkit PR https://github.com/thodan/bop_toolkit/pull/82 is merged.
