### 2023-02-03 66be77e
fix(#810): remove leftover render.__init__
