### 2023-02-05 f398f93
fix(paper): Corrects "open source renderer" row in table 1
