### 2023-02-09 0f69058
fix(#815): Add closing the pipe handles
