### 2023-02-17 71b948a
feat(set_denoiser): add openimagedenoise denoiser close #818
