+++ blenderproc/python/renderer/RendererUtility.py
@@ -28,7 +28,7 @@ def set_denoiser(denoiser: Optional[str]):
 
     Automatically disables all previously activated denoiser.
 
-    :param denoiser: The name of the denoiser which should be enabled. Options are "INTEL", "OPTIX" and None. \
+    :param denoiser: The name of the denoiser which should be enabled. Options are "INTEL", "OPTIX", "OPENIMAGEDENOISE" and None. \
                      If None is given, then no denoiser will be active.
     """
     # Make sure there is no denoiser active
@@ -39,6 +39,10 @@ def set_denoiser(denoiser: Optional[str]):
         bpy.context.scene.cycles.use_denoising = True
         bpy.context.view_layer.cycles.use_denoising = True
         bpy.context.scene.cycles.denoiser = "OPTIX"
+    elif denoiser.upper() == "OPENIMAGEDENOISE":
+        bpy.context.scene.cycles.use_denoising = True
+        bpy.context.view_layer.cycles.use_denoising = True
+        bpy.context.scene.cycles.denoiser = "OPENIMAGEDENOISE"
     elif denoiser.upper() == "INTEL":
         # The intel denoiser is activated via the compositor
         bpy.context.scene.use_nodes = True
