### 2023-02-18 #4901 @Kevin-Mattheus-Moerman
@themasterlink  @cornerfarmer based on the above positive reviews, I am happy to process this work for acceptance in JOSS. To proceed we require your help to work on the following:

- [ ] Please archive a copy of the software on ZENODO and report back here with the DOI. You can do this manually or you can follow these automated steps: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

Please check the ZENODO archive and check the following (you may need to manually edit the ZENODO entry for this):
- [ ] The  title should match the paper title
- [ ] The author set and order should be the same. You may also (recommended) add ORCID's for each author at this point.
- [ ] The license listed should match your software license
- [ ] This review is currently defined as for version `v2.4.1`. During the review the version may have been incremented (and some alter it when they create the archive). The version listed for this review and for the ZENODO archive should match. Furthermore this version tag should appear as a tagged release on your GitHub repository. Using the above automated steps ZENODO will match the version listed on GitHub automatically. But do check this and alter it if needed. Next report this updated version tag to me and I can update the one listed for this review issue.

I have read the paper and it seems in order to me. If you wanted to proofread it yourself once more, now is the time to do so. In particular check that the author set and names, and the acknowledgements are accurate.
